Adding custom process generators to compiler
